backend/requirements.txt
Django==3.2.25
djangorestframework==3.15.1
django-cors-headers==4.1.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
Pillow==10.0.1
django-filter==23.5

# Azure Production Dependencies
gunicorn==21.2.0
whitenoise==6.6.0
opencensus-ext-django==0.8.0
opencensus-ext-logging==0.1.0

# Additional Production Dependencies
django-storages==1.14.2
azure-storage-blob==12.19.0
